+++ lumin/nn/models/blocks/head.py
@@ -1,79 +1,78 @@
 import numpy as np
-from typing import Dict, Optional, List, Callable
+from typing import Dict, Optional, Callable, List
 from glob import glob
 from collections import OrderedDict
 from pathlib import Path
 import os
+from abc import abstractmethod
 
 import torch.nn as nn
 from torch.tensor import Tensor
 import torch
 
-from ..helpers import Embedder
+from ..helpers import CatEmbedder
 from ..initialisations import lookup_normal_init
-from ..layers.activations import lookup_act
 from ....plotting.plot_settings import PlotSettings
 from ....plotting.interpretation import plot_embedding
+from .abs_block import AbsBlock
 
 
-class CatEmbHead(nn.Module):
+class AbsHead(AbsBlock):
+    def __init__(self, cont_feats:List[str], cat_embedder:Optional[CatEmbedder]=None, 
+                 lookup_init:Callable[[str,Optional[int],Optional[int]],Callable[[Tensor],None]]=lookup_normal_init, freeze:bool=False):
+        super().__init__(lookup_init=lookup_init, freeze=freeze)
+        self.cont_feats,self.cat_embedder = cont_feats,cat_embedder
+        self.n_cont_in = len(cont_feats)
+
+    @abstractmethod
+    def _map_outputs(self) -> Dict[str,List[int]]: pass
+
+
+class CatEmbHead(AbsHead):
     r'''
     Standard model head for columnar data.
     Provides inputs for continuous features and embedding matrices for categorical inputs, and uses a dense layer to upscale to width of network body.
     Designed to be passed as a 'head' to :class:ModelBuilder.
     Supports batch normalisation and dropout (at separate rates for pre-dense continuous, pre-dense categorical embeddings, and post dense).
-    Continuous features are expected to be the first n_cont_in columns of input tensors and categorical features the remaining columns.
-    Embedding arguments for categorical features are set using a :class:Embedder.
+    Continuous features are expected to be the first len(cont_feats) columns of input tensors and categorical features the remaining columns.
+    Embedding arguments for categorical features are set using a :class:CatEmbedder.
 
     Arguments:
-        n_cont_in: number of continuous inputs to expect
-        act: string representation of argument to pass to lookup_act
-        do: if not None will add a dropout layer with dropout rate do after the dense layer
+        cont_feats: list of names of continuous input features
         do_cont: if not None will add a dropout layer with dropout rate do acting on the continuous inputs prior to concatination wih the categorical embeddings
         do_cat: if not None will add a dropout layer with dropout rate do acting on the categorical embeddings prior to concatination wih the continuous inputs
-        bn: whether to add a batch normalisation layer after the activation function
-        cat_embedder: :class:Embedder providing details of how to embed categorical inputs
+        cat_embedder: :class:CatEmbedder providing details of how to embed categorical inputs
         lookup_init: function taking choice of activation function, number of inputs, and number of outputs an returning a function to initialise layer weights.
-        lookup_act: function taking choice of activation function and returning an activation function layer
         freeze: whether to start with module parameters set to untrainable
 
     Examples::
-        >>> head = CatEmbHead(n_cont_in=30, n_out=100, act='relu')
-        >>> head = CatEmbHead(n_cont_in=25, n_out=100, act='relu',  cat_embedder=Embedder.from_fy(train_fy))
-        >>> head = CatEmbHead(n_cont_in=25, n_out=100, act='swish', cat_embedder=Embedder.from_fy(train_fy), do=0.1)
-        >>> head = CatEmbHead(n_cont_in=25, n_out=100, act='prelu', cat_embedder=Embedder.from_fy(train_fy), bn=True, lookup_init=lookup_uniform_init)
+        >>> head = CatEmbHead(cont_feats=cont_feats)
+        >>> head = CatEmbHead(cont_feats=cont_feats, cat_embedder=CatEmbedder.from_fy(train_fy))
+        >>> head = CatEmbHead(cont_feats=cont_feats, cat_embedder=CatEmbedder.from_fy(train_fy), do_cont=0.1, do_cat=0.05)
+        >>> head = CatEmbHead(cont_feats=cont_feats, cat_embedder=CatEmbedder.from_fy(train_fy), lookup_init=lookup_uniform_init)
     '''
 
-    def __init__(self, n_cont_in:int, act:str, do:float, do_cont:float, do_cat:float, bn:bool, cat_embedder:Optional[Embedder]=None, 
-                 lookup_init:Callable[[str,Optional[int],Optional[int]],Callable[[Tensor],None]]=lookup_normal_init,
-                 lookup_act:Callable[[str],nn.Module]=lookup_act, freeze:bool=False):
-        super().__init__()
-        self.n_cont_in,self.do_cont,self.do_cat,self.bn,self.cat_embedder = n_cont_in,do_cont,do_cat,bn,cat_embedder
-        self.act,self.lookup_init,self.lookup_act,self.freeze = act,lookup_init,lookup_act,freeze
-        if self.cat_embedder is None: self.cat_embedder = Embedder([], [])
-        if self.cat_embedder.n_cat_in > 0: 
+    def __init__(self, cont_feats:List[str], do_cont:float=0, do_cat:float=0, cat_embedder:Optional[CatEmbedder]=None, 
+                 lookup_init:Callable[[str,Optional[int],Optional[int]],Callable[[Tensor],None]]=lookup_normal_init, freeze:bool=False):
+        super().__init__(cont_feats=cont_feats, cat_embedder=cat_embedder, lookup_init=lookup_init, freeze=freeze)
+        self.do_cont,self.do_cat, = do_cont,do_cat
+        if self.cat_embedder is None: self.cat_embedder = CatEmbedder([], [])
+        if self.cat_embedder.n_cat_in > 0:
             self.embeds = nn.ModuleList([nn.Embedding(ni, no) for _, ni, no in self.cat_embedder])
             if self.cat_embedder.emb_load_path is not None: self._load_embeds()
             if self.do_cat   > 0: self.emb_do     = nn.Dropout(self.do_cat)
         self.n_out = self.n_cont_in if self.cat_embedder.n_cat_in == 0 else self.n_cont_in+np.sum(self.cat_embedder.emb_szs)
         if self.do_cont  > 0: self.cont_in_do = nn.Dropout(self.do_cont)
         if self.freeze: self.freeze_layers()
-    
-    def __getitem__(self, key:int) -> nn.Module: return self.layers[key]
+        self._map_outputs()
 
-    def freeze_layers(self):
-        r'''
-        Make parameters untrainable
-        '''
-
-        for p in self.parameters(): p.requires_grad = False
-    
-    def unfreeze_layers(self):
-        r'''
-        Make parameters trainable
-        '''
-
-        for p in self.parameters(): p.requires_grad = True
+    def _map_outputs(self) -> Dict[str,List[int]]:
+        self.feat_map = {}
+        for i, f in enumerate(self.cont_feats): self.feat_map[f] = [i]
+        offset = self.n_cont_in
+        for f, _, sz in self.cat_embedder:
+            self.feat_map[f] = list(range(offset, offset+sz))
+            offset += sz
         
     def forward(self, x_in:Tensor) -> Tensor:
         if self.cat_embedder.n_cat_in > 0:
@@ -83,7 +82,7 @@ class CatEmbHead(nn.Module):
         if self.n_cont_in > 0:
             x_cont = x_in[:,:self.n_cont_in]
             if self.do_cont > 0: x_cont = self.cont_in_do(x_cont) 
-            x = torch.cat((x, x_cont), dim=1) if self.cat_embedder.n_cat_in > 0 else x_cont
+            x = torch.cat((x_cont, x), dim=1) if self.cat_embedder.n_cat_in > 0 else x_cont
         return x
     
     def _load_embeds(self, path:Optional[Path]=None) -> None: